requirements.txt
aiogram>=3.4.0
supabase>=2.4.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
